backend/requirements.txt
fastapi[standard]
peewee
psycopg2
pymysql
requests
peewee-migrations
pandas
debugpy
sqlalchemy
pyjwt
passlib
bcrypt